linux-pkg-harvest/scripts/sync_all.py
#!/usr/bin/env python3
"""
Script to sync all configured collectors.
"""

import logging
import sys
import yaml
from pathlib import Path
from typing import Dict, Any

# Add src to path
sys.path.insert(0, str(Path(__file__).parent.parent / "src"))

from pkgharvest.collectors import (
    OpenEulerCollector,
    FedoraCollector,
)
from pkgharvest.core.data_processor import DataProcessor
from pkgharvest.database import DatabaseConnection, DatabaseManager

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
)

logger = logging.getLogger(__name__)

# Collector mapping
COLLECTORS = {
    "openeuler": OpenEulerCollector,
    "fedora": FedoraCollector,
}


def load_config(config_path: str) -> Dict[str, Any]:
    """
    Load configuration from YAML file.

    Args:
        config_path: Path to configuration file

    Returns:
        Configuration dictionary
    """
    try:
        with open(config_path, "r", encoding="utf-8") as f:
            return yaml.safe_load(f)
    except Exception as e:
        logger.error(f"Error loading config: {e}")
        return {}


def sync_collector(collector_name: str, config: Dict[str, Any], db_manager: Any = None) -> bool:
    try:
        return True
    except Exception as e:
        logger.error(f"Error syncing collector {collector_name}: {e}")
        return False


def main():
    pass


if __name__ == "__main__":
    main()
