linux-pkg-harvest/requirements.txt
requests>=2.31.0
urllib3>=2.0.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
packaging>=23.0
pymysql>=1.1.0
sqlalchemy>=2.0.0

